[Feat/#36] 질문카테고리 UI CategoryUtil로 따로 확장함수 로직 빼기
